updated SLF4J dependency to latest modularized version (2.0.0.alpha0)
