+++ libraries/pi4j-library-pigpio/src/test/java/com/pi4j/library/pigpio/test/pwm/TestSoftwarePwmUsingTestHarness.java
@@ -51,7 +51,7 @@ public class TestSoftwarePwmUsingTestHarness {
 
     @BeforeAll
     public static void initialize() {
-        //System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        //System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         System.out.println();
         System.out.println("************************************************************************");
