+++ libraries/pi4j-library-pigpio/src/test/java/com/pi4j/library/pigpio/test/spi/TestSpiUsingTestHarness.java
@@ -55,7 +55,7 @@ public class TestSpiUsingTestHarness {
 
     @BeforeAll
     public static void initialize() {
-        //System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        //System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         System.out.println();
         System.out.println("************************************************************************");
