+++ pi4j-example/src/main/java/com/pi4j/example/gpio/digital/DigitalInputExample.java
@@ -60,7 +60,7 @@ public class DigitalInputExample {
      */
     public static void main(String[] args) throws Exception {
         // configure logging output
-        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
+        System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
 
         // TODO :: REMOVE TEMPORARY PROPERTIES WHEN NATIVE PIGPIO LIB IS READY
         // this temporary property is used to tell
