+++ pi4j-example/src/main/java/com/pi4j/example/gpio/digital/DigitalInputExampleUsingDependencyInjection.java
@@ -55,7 +55,7 @@ public class DigitalInputExampleUsingDependencyInjection {
     public static void main(String[] args) throws Exception {
 
         // configure logging output
-        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         // TODO :: REMOVE TEMPORARY PROPERTIES WHEN NATIVE PIGPIO LIB IS READY
         // this temporary property is used to tell
