+++ pi4j-example/src/main/java/com/pi4j/example/i2c/I2cRawDeviceExample.java
@@ -53,7 +53,7 @@ public class I2cRawDeviceExample {
      * @throws java.lang.Exception if any.
      */
     public static void main(String[] args) throws Exception {
-        //System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        //System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         // TODO :: REMOVE TEMPORARY PROPERTIES WHEN NATIVE PIGPIO LIB IS READY
         // this temporary property is used to tell
