+++ pi4j-test/src/main/java/com/pi4j/test/Main.java
@@ -55,7 +55,7 @@ public class Main {
     public static void main(String[] args) throws Exception {
 
         // set logging
-        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        //System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         // Initialize Pi4J with an auto context
         // An auto context includes AUTO-DETECT BINDINGS enabled
