+++ pi4j-test/src/test/java/com/pi4j/test/annotations/DependencyInjectionRegisterCustomPlatformTest.java
@@ -42,7 +42,7 @@ public class DependencyInjectionRegisterCustomPlatformTest {
 //
 //    @Before
 //    public void beforeTest() throws Pi4JException {
-//        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
+//        System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
 //
 //        // Initialize Pi4J with an empty context
 //        // An empty context disables AUTO-DETECT loading
