+++ pi4j-test/src/test/java/com/pi4j/test/annotations/DependencyInjectionRegisterCustomProviderTest.java
@@ -46,7 +46,7 @@ public class DependencyInjectionRegisterCustomProviderTest {
 //
 //    @Before
 //    public void beforeTest() throws Pi4JException {
-//        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
+//        System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
 //
 //        // Initialize Pi4J with an empty context
 //        // An empty context disables AUTO-DETECT loading
