+++ pi4j-test/src/test/java/com/pi4j/test/annotations/DependencyInjectionRegistryTest.java
@@ -49,7 +49,7 @@ public class DependencyInjectionRegistryTest {
     @Before
     public void beforeTest() throws Pi4JException {
 
-        System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
+        System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
 
         // initialize Pi4J with an auto context
         // An auto context includes AUTO-DETECT BINDINGS enabled
