+++ plugins/pi4j-plugin-pigpio/src/test/java/com/pi4j/plugin/pigpio/test/spi/TestSpiUsingTestHarness.java
@@ -1,11 +1,11 @@
-package com.pi4j.plugin.pigpio.serial;
+package com.pi4j.plugin.pigpio.test.spi;
 
 /*-
  * #%L
  * **********************************************************************
  * ORGANIZATION  :  Pi4J
  * PROJECT       :  Pi4J :: PLUGIN   :: PIGPIO I/O Providers
- * FILENAME      :  TestSerialUsingTestHarness.java
+ * FILENAME      :  TestSpiUsingTestHarness.java
  *
  * This file is part of the Pi4J project. More information about
  * this project can be found here:  https://pi4j.com/
@@ -33,12 +33,11 @@ import com.pi4j.Pi4J;
 import com.pi4j.context.Context;
 import com.pi4j.exception.LifecycleException;
 import com.pi4j.exception.Pi4JException;
-import com.pi4j.io.serial.Baud;
-import com.pi4j.io.serial.Serial;
+import com.pi4j.io.spi.Spi;
 import com.pi4j.library.pigpio.PiGpio;
-import com.pi4j.plugin.pigpio.TestEnv;
-import com.pi4j.plugin.pigpio.provider.serial.PiGpioSerialProvider;
-import com.pi4j.plugin.pigpio.provider.serial.PiGpioSerialProviderImpl;
+import com.pi4j.plugin.pigpio.test.TestEnv;
+import com.pi4j.plugin.pigpio.provider.spi.PiGpioSpiProvider;
+import com.pi4j.plugin.pigpio.provider.spi.PiGpioSpiProviderImpl;
 import com.pi4j.test.harness.ArduinoTestHarness;
 import com.pi4j.test.harness.TestHarnessInfo;
 import com.pi4j.test.harness.TestHarnessPins;
@@ -51,13 +50,13 @@ import java.nio.ByteBuffer;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-@DisplayName("PIGPIO Plugin :: Test Serial Communication using Test Harness")
+@DisplayName("PIGPIO Plugin :: Test SPI Communication using Test Harness")
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class TestSerialUsingTestHarness {
+public class TestSpiUsingTestHarness {
 
-    private static String SERIAL_DEVICE = "/dev/ttyS0";
-    private static int BAUD_RATE = Baud._38400.value();
-    private static int TEST_HARNESS_UART = 3;
+    private static int SPI_CHANNEL = 0;
+    private static int BAUD_RATE = 100000;
+    private static int TEST_HARNESS_SPI_CHANNEL = 10;
 
     private static byte SAMPLE_BYTE = 0x0d;
     private static int SAMPLE_WORD = 0xFFFA;
@@ -68,16 +67,16 @@ public class TestSerialUsingTestHarness {
 
     private static PiGpio piGpio;
     private static Context pi4j;
-    private static Serial serial;
+    private static Spi spi;
 
     @BeforeAll
     public static void initialize() {
         // configure logging output
-        //System.setProperty(org.slf4j.impl.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
+        //System.setProperty(org.slf4j.simple.SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "TRACE");
 
         System.out.println();
         System.out.println("************************************************************************");
-        System.out.println("INITIALIZE TEST (" + TestSerialUsingTestHarness.class.getName() + ")");
+        System.out.println("INITIALIZE TEST (" + TestSpiUsingTestHarness.class.getName() + ")");
         System.out.println("************************************************************************");
         System.out.println();
 
@@ -103,22 +102,21 @@ public class TestSerialUsingTestHarness {
             System.out.println();
             System.out.println("RESET ALL PINS ON TEST HARNESS; (" + reset.total + " pin reset)");
 
-            // enable the Serial Echo (Loopback) function on the test harness for these tests
-            harness.enableSerialEcho(TEST_HARNESS_UART,  BAUD_RATE);
+            // enable the SPI Echo (Loopback) function on the test harness for these tests
+            harness.enableSpiEcho(TEST_HARNESS_SPI_CHANNEL);
             System.out.println();
-            System.out.println("ENABLE SERIAL UART [" + TEST_HARNESS_UART + "] ON TEST HARNESS; BAUD=" + BAUD_RATE);
+            System.out.println("ENABLE SPI CHANNEL [" + TEST_HARNESS_SPI_CHANNEL + "] ON TEST HARNESS; BAUD=" + BAUD_RATE);
 
             // close connection to test harness
             harness.close();
 
-            // create SERIAL config
-            var config  = Serial.newConfigBuilder(pi4j)
-                    .id("my-serial-port")
-                    .name("My Serial Port")
-                    .device(SERIAL_DEVICE)
-                    .baud8N1(BAUD_RATE)
+            // create SPI config
+            var config  = Spi.newConfigBuilder(pi4j)
+                    .id("my-spi")
+                    .name("My SPI")
+                    .address(SPI_CHANNEL)
+                    .baud(BAUD_RATE)
                     .build();
-
 
             // TODO :: THIS WILL NEED TO CHANGE WHEN NATIVE PIGPIO SUPPORT IS ADDED
             piGpio = TestEnv.createPiGpio();
@@ -126,15 +124,15 @@ public class TestSerialUsingTestHarness {
             // initialize the PiGpio library
             piGpio.initialize();
 
-            // create SERIAL provider instance to test with
-            PiGpioSerialProvider provider = new PiGpioSerialProviderImpl(piGpio);
+
+            // create SPI provider instance to test with
+            PiGpioSpiProvider provider = new PiGpioSpiProviderImpl(piGpio);
 
             // initialize Pi4J instance with this single provider
             pi4j = Pi4J.newContextBuilder().add(provider).build();
 
-            // create serial instance
-            serial = pi4j.create(config);
-
+            // create SPI instance
+            spi = pi4j.create(config);
         } catch (IOException e){
             e.printStackTrace();
         } catch (Pi4JException e) {
@@ -148,13 +146,12 @@ public class TestSerialUsingTestHarness {
     public static void terminate() throws IOException, LifecycleException {
         System.out.println();
         System.out.println("************************************************************************");
-        System.out.println("TERMINATE TEST (" + TestSerialUsingTestHarness.class.getName() + ") ");
+        System.out.println("TERMINATE TEST (" + TestSpiUsingTestHarness.class.getName() + ") ");
         System.out.println("************************************************************************");
         System.out.println();
 
-        // close down serial
-        //if(serial.isOpen())
-        serial.close();;
+        // close down SPI channel
+        spi.close();;
 
         // shutdown the PiGpio library
         piGpio.shutdown();
@@ -173,81 +170,80 @@ public class TestSerialUsingTestHarness {
     }
 
     @Test
-    @DisplayName("SERIAL :: Verify SERIAL Instance")
+    @DisplayName("SPI :: Verify SPI Instance")
     @Order(1)
-    public void testSerialInstance() throws Exception {
-        // ensure that the SERIAL instance is not null;
-        assertNotNull(serial);
+    public void testSpiInstance() throws Exception {
+        // ensure that the SPI instance is not null;
+        assertNotNull(spi);
 
         // ensure connection is open
-        assertTrue("SERIAL INSTANCE IS NOT OPEN", serial.isOpen());
+        assertTrue("SPI INSTANCE IS NOT OPEN", spi.isOpen());
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE (WRITE)")
+    @DisplayName("SPI :: Test BYTE (WRITE)")
     @Order(2)
-    public void testSerialSingleByteWrite() throws Exception {
-        // write a single byte to the serial device
-        serial.write(SAMPLE_BYTE);
+    public void testSpiSingleByteWrite() throws Exception {
+        // write a single byte to the SPI device
+        spi.write(SAMPLE_BYTE);
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE (READ)")
+    @DisplayName("SPI :: Test BYTE (READ)")
     @Order(3)
-    public void testSerialSingleByteRead() throws Exception {
-        // read single byte from the serial device
-        Assert.assertEquals(SAMPLE_BYTE, serial.readByte());
+    public void testSpiSingleByteRead() throws Exception {
+        // read single byte from the SPI device
+        Assert.assertEquals(SAMPLE_BYTE, spi.readByte());
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE-ARRAY (WRITE)")
+    @DisplayName("SPI :: Test BYTE-ARRAY (WRITE)")
     @Order(4)
-    public void testSerialByteArrayWrite() throws Exception {
-        // write an array of data bytes to the serial device
-        serial.write(SAMPLE_BYTE_ARRAY);
+    public void testSpiByteArrayWrite() throws Exception {
+        // write an array of data bytes to the SPI device
+        spi.write(SAMPLE_BYTE_ARRAY);
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE-ARRAY (READ)")
+    @DisplayName("SPI :: Test BYTE-ARRAY (READ)")
     @Order(5)
-    public void testSerialByteArrayRead() throws Exception {
-        // read an array of data bytes from the serial device
-        byte[] byteArray = serial.readNBytes(SAMPLE_BYTE_ARRAY.length);
-        Assert.assertArrayEquals(SAMPLE_BYTE_ARRAY, byteArray);
+    public void testSpiByteArrayRead() throws Exception {
+        // read an array of data bytes from the SPI device
+        byte[] byteArray = spi.readNBytes(SAMPLE_BYTE_ARRAY.length);
+        Assert.assertEquals(SAMPLE_BYTE_ARRAY[SAMPLE_BYTE_ARRAY.length-1], byteArray[0]);
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE-BUFFER (WRITE)")
+    @DisplayName("SPI :: Test BYTE-BUFFER (WRITE)")
     @Order(6)
-    public void testSerialByteBufferWrite() throws Exception {
-        // write a buffer of data bytes to the serial device
-        serial.write(SAMPLE_BUFFER);
+    public void testSpiByteBufferWrite() throws Exception {
+        // write a buffer of data bytes to the SPI device
+        spi.write(SAMPLE_BUFFER);
     }
 
     @Test
-    @DisplayName("SERIAL :: Test BYTE-BUFFER (READ)")
+    @DisplayName("SPI :: Test BYTE-BUFFER (READ)")
     @Order(7)
-    public void testSerialByteBufferRead() throws Exception {
-        // read a buffer of data bytes from the serial device
-        ByteBuffer buffer = serial.readByteBuffer(SAMPLE_BUFFER.capacity());
-        Assert.assertArrayEquals(SAMPLE_BUFFER.array(), buffer.array());
+    public void testSpiByteBufferRead() throws Exception {
+        // read a buffer of data bytes from the SPI device
+        ByteBuffer buffer = spi.readByteBuffer(SAMPLE_BUFFER.capacity());
+        Assert.assertEquals(SAMPLE_BUFFER.get(SAMPLE_BUFFER.capacity()-1), buffer.get(0));
     }
 
     @Test
-    @DisplayName("SERIAL :: Test ASCII STRING (WRITE)")
+    @DisplayName("SPI :: Test ASCII STRING (WRITE)")
     @Order(8)
-    public void testSerialAsciiStringWrite() throws Exception {
-        // write a string of data to the serial device
-        serial.write(SAMPLE_STRING);
+    public void testSpiAsciiStringWrite() throws Exception {
+        // write a string of data to the SPI device
+        spi.write(SAMPLE_STRING);
     }
 
     @Test
-    @DisplayName("SERIAL :: Test ASCII STRING (READ)")
+    @DisplayName("SPI :: Test ASCII STRING (READ)")
     @Order(9)
-    public void testSerialAsciiStringRead() throws Exception {
-        // read a string of data from the serial device
-        String testString = serial.readString(SAMPLE_STRING.length());
-        Assert.assertEquals(SAMPLE_STRING, testString);
+    public void testSpiAsciiStringRead() throws Exception {
+        // read a string of data from the SPI device
+        String testString = spi.readString(SAMPLE_STRING.length());
+        Assert.assertEquals(SAMPLE_STRING.substring(SAMPLE_STRING.length()-1), testString.substring(0, 1));
     }
-
 }
