+++ pom.xml
@@ -204,7 +204,7 @@
         <junit.version>4.12</junit.version>
         <junit.jupiter.version>5.5.1</junit.jupiter.version>
         <junit.platform.version>1.5.1</junit.platform.version>
-        <slf4j.version>1.7.26</slf4j.version>
+        <slf4j.version>2.0.0-alpha0</slf4j.version>
         <jSerialComm.version>2.5.1</jSerialComm.version>
         <gson.version>2.8.5</gson.version>
         <json.version>20190722</json.version>
